Register typedefs under --enable-intel-vector-types in C++
In C++ (not in C), Mercurium ignores redundant typedefs like

typedef struct A { ... } A;

This is usually not a problem, but for Intel C/C++ there is a typedef
for the vectorial types __m{128,256,512}{,d,i}. That typedef-name is
used by the fronted to determine builtin calls. Not emitting a typedef
causes the frontend to emit a usual extern call (albeit without warning).
This causes the link step to fail.

We are not interested for C++ to keep these redundant typedefs for the
moment, so we manually register these typedefs and we prepend explicit
declarations at the beginning of the tree.

This commit fixes #2186