Fix Liveness Analysis for tasks
This analysis is now more conservative in the sense that: if there
is a possible path where a variable is alive, then we keep it alive.

This commit changes the directionality of the data-flow algorithm from
top-bottom to bottom-top, this way the algorithm will iterate less times.
This change has broken the test 'defs_01.c'. This is because return
statements in the middle of the code cause some graph nodes to have
their exit nodes unconnected.
Because of that, the assumed condition of having all graph nodes with
an entry node at the beginning and an exit node at the end does not
fulfill.

This commit fixes the test that was broken with the temporary commit
78c04e6e7d4cf26f279ac068a03e2f8a48e121a3.