Fixes in CFG Analysis
Due to transforming Liveness analysis in a bottom-up analysis,
a problem in the structure of the PCFG has appeared: the occurrence
of return statements in the middle of the code causes their container
graph node to have their Exit node unconnected. This breaks the assumed
rule that all graph nodes start with an Entry node and finish with
an Exit node. This commit fixes this problem and also fixes the method
that cleans up the PCFG from bottom-up to handles this extrange situation.

This commit fixes test defs_01.c, broken in commit
6a7a083d784e6a7a4c8769eee3147999da186c74.