Add temporal patch to overlap algorithm
A list of IV increments is necessary to apply the transformation when
unrolling is required. We don't have time now to implement it so
this is a temporal patch that makes the transformation work only
under certain circumstances. If they are not satisfied, the overlap
algorithm emits an error.

Let's keep track of this in the ticket related with the IV.

See #2198