UseDef for unreachable code functions refactoring
UseDef analysis registers at the begining the c/c++ library
methods specified in the data base (accordingly to the language
of the files being compiled) and their usage.
When a function whose code is not reachable is found, we just
look up the function in the previously loaded data base and
propagate the usage information.

See #2189