Avoid reporting unknown race conditions
Array subscripts may not cause race conditions when some of the subscripts
are defined between the creation of the two tasks. This commit implements
all cases (dependencies between two instances of the same task and dependencies
between two different tasks). We add a test case (mandelbrot).
See #2167

This commit also allows defining empty list in the assertion clauses
to check when the Analysis reports no variable for a given analysis.
See #2129