Several fixes to use_def
Because of turning analysis check phase more restrictive, some errors
in the use-def analysis implementation showed up.

The fixed errors are related to:
* local variables usage was propagated to the whole function usage:
this is wrong when propagating this usage to a possible calling function.
* default arguments: IPA transformation must avoid the nodecl DefaultArgument
because it is not an expression and shall not be treated as such.

Now, only those correctness clauses that are explicitely specified
in the test cases are checked (previously one clause realated to each
analysis triggered checking all possible clauses related to that analysis.