Minor fixes to correctness messages
Variables data-sharing: when a variable is reported as "dead", then
it is not reported as an incoherent firstprivate or an incoherent private,
because the messages are contradictory.

Fix lint for variables scope which are not used within the task:
now, we check for sub-objects as well as for structurally equal nodecls.

See #2167