Fix TDG JSON dependency expressions
When more than one pair of memory accesses may cause a dependency to
occur, then we net to use an Or operator to concatenate them, instead
of using an And operator, because only one of them fulfilling will
cause the dependency to exist.

This commit also changes the values of the variables involved in a
dependency expressions. When it is an induction variable, the value is
reported as an offset since the iteration starts. Otherwise, values
from range analysis are reported.

Fixes #2217