Fix mapping of variables during deep-copies to be deterministic
Rather than traversing names and using them to register symbols,
traverse symbols first and create them but do not insert to the scope.

Then traverse again and register the names.

Also we change the map lookup to confuse ourselves with externally
mapped symbols.

This makes the process deterministic.

This commit fixes #2234