Fix comparing analysis nodecl maps
Reaching Definitions Analysis uses a multimap whose order is non-deterministic.
This means that two entries for the same variable are not necessarily ordered
always the same way because only the kind, the symbol and the constant are checked
when a pair is inserted.
We take into account this inderteminism when comparing two maps.