Make sure we can transitively load OmpSs task information
The information of OmpSs tasks is not propagated but we can attempt
to load the module where the task was declared.

Sadly, when doing this we observed that dummy arguments were duplicated
which led to incorrect behaviour.

Now, likewise we do for symbols from (partially) loaded modules,
we do not create new dummy arguments if the function it belongs
has already been loaded (i.e. indirectly).

This commit fixes #2233