Tag more nodes as constant values
We can tag nodecl trees with a const_value_t and it means that we know
this will be its constant value (ahead of time).

C/C++ frontends equaled the presence of the tag with the concept of
"constant expression" in the C/C++ Standard. The Standard is actually
rather restrictive, since cases like

   const int a[2] = {1, 2};

neither a, a[0] or a[1] are considered constant expressions but
given that it is undefined behaviour to modify constant objects,
it should be sensible to assume that their known values will be
{1,2}, 1 and 2 respectively. (Note that in C++11, if the array above
were declared as constexpr, all three expressions would be
constant).

So we decouple the Standard constant expression property and the
presence of the constant value tag. It has been necessary to add
some extra semantic checks to avoid accepting expressions with
constant values that happen to be not constant in the Standard.

This extended tagging of expressions was required for analysis and
SIMD transformations, that will benefit from this. See ticket #2237

For testing whether an expression has a const-value, a new
@const-value-check@ extension has been aded. See
03_extensions_c.dg/success_const_value_check_01.c for an example
on how to use it in tests.

This commit fixes #2239

This commit also fixes #2240. Rather than disabling full constant
evaluation in non-strict operators (&& and ||) we just disable constexpr
function calls in this case. This change may only affect C++11
and will get back some missing const values because of this.

A workaround was added due to #2240 that should not be
needed anymore. See #2241

Due to union initializations, there is a new "unknown" const value.  It
expresses the fact that we know the entity is constant (the entity would
end in the constant data part of the program) but we cannot come up with
a value. Currently this only happens for unions: the inactive fields
(the ones not initialized) obviously have some constant value but which
one is, is unknown. It might happen that such "value" is not even
representng a valid value for that object (i.e. consider a union with an
int and a float field, serveral int values map to NaN). In a future this
new const value kind might be used in an hypothetical constant
propagation phase that propagates structured values that may include
unknown values:

   struct A { int x; int y; };

   void f(struct A a) { if (a.x < 3) { maybe_dead_code(); } };

   void g(int y) { struct A a; a.y = y; a.x = 3; f(a); }

in the example above, the value of a.y may be left unknown when
propagating the constant in the call to 'f'.