Enhance replacement for nodecls other than symbol
Different parts of the analysis use a nodecl replacer that was implemented
only to replace Symbol by Nodecl. This transformation is necessary when
we want to replace the symbols of the parameters of a function with their
corresponding arguments.
Other than that, we need to replace any kind of memory access (Symbol,
ClassMemberAccess and ArraySubscript) by a Nodecl.