Add ObjectInit to false IV detection
Although we check local variables to purge the IV list
we also want to implement the ObjectInit to cover all possibilities.

Enhance IV checker: now the sets of asserted IV and computed IV must
be structurallty equal.

This commit also fixes a mistake in the dataref_contains_dataref method,
for ClassMemberAccess nodecls.

See #2253