Implement basic support for derived IV recognition
This commit transforms the induction variable analysis into a
convergent algorithm which, in the first iteration recognizes
the Basic IVs and, in the next iterations while something changes,
recognizes the Derives IVs.

It cannot be completed because the current way to discover
the boundaries of an IV is too simple (based on the reaching
definitions and the loop control). So, we are thinking on using
Range Analysis. We'll try that option in an experimental branch.

See #1708