Change the logic of C++11 pack declarators
The existing logic was poor and was not able to cover many cases
where using {function,template}-parameter-packs is legitimate.

This commit reworks the strategy, adding a bit of state to the build
scope routines. This is undesirable but is unavoidable given that
a {function,template} parameter pack is allowed as long as it appears
inside a pack expansion, and this may happen more or less arbitrarily
in the syntax of declarators.

This commit fixes #2271