Extend the task description in intrumentation mode
The TDG plugin changes in a way that all tasks with the same label
will be depicted with the same color. This forces us to distinguish
those descriptions which contain a label and those which does not.

The new format of the description is:
func_type|func_label @ file @ line @ name_type
, where name_type is "LABEL" or "FUNCTION"

This is related to the nanos ticket #1037.