Make sure do_koenig_lookup knows about auto
In [62cfecbca8253540c513bd2dc02bae73fa9c9fd7] we made `auto` a
non-dependent type. This is caused by C++14 requiring extra deductions
that might be skipped if auto was considered dependent.

An expression involving a name that is still `auto` is considered
dependent, but the type itself (i.e. the type of a symbol like `g`
above) needs an extra check.

The real issue happens in `do_koenig_lookup` which overlaps a bit
with `cxx_compute_name_from_entry_list` due to the very special process
required by calls to unqualified names.

This commit fixes #2289