Fix IV analysis for IVs with non-const increment
The Inductin Variable analysis was assuming that variables suitable
to be Induction Variables ( iv += ...; iv = iv + ...; etc.) had always
an increment with a constant value associated. This may not be true and,
in that case, the analysis fails.
This commit protects the accesses to Nodecl::NodeclBase::get_constant()
with a call to Nodecl::NodeclBase::is_constant().

This commit also enhances the performance of IVs analysis by storing
those variables that already have been discarded as IVs, so they are
not checked again if they appear in multiple statements.

Fixes #2295