Improve KNC prefetcher
Prefetching instructions now are placed (in order or appearance) at
the beginning of the loop body. This seems to be better for the icc's
instruction scheduler.

Adjacent vloads (a[i], a[i+16]) now produce only one pref instruction.
Now, the overlap algorithm optionally considers adjacent vloads as
part of the same overlap group.

See #2297