Fix C99 inline semantics by forcing extern
Note that in general we do not explicitly emit extern for function
declarations that have external linkage. In this case though, omitting
extern can change the behaviour of inline functions, so we need
to emit extern.

Note that this makes Mercurium incompatible with GNU C90 inline semantics.
Given that GCC 5.0 will, according to the docs, default to -std=c11
(C2011) we do not feel compelled to make special cases for GNU C90.

This commit fixes #2293