Rework dependent array-sections nodes
Remove nodes Nodecl::CxxArraySectionRange and Nodecl::CxxArraySectionSize
that were used for dependent array-sections. Instead we will use a plain
Nodecl::ArraySubscript (where the subscript will contain ranges as
usual).

By removing these two trees we can leverage on partial dependent types
(like "pointer to <dependent-thing>") that lets us to apply transformations
in OmpSs that only care about the structure of the type.

So

  T t;
  #pragma omp task inout(t[1:2])
  { }

is not supported because 't' is unknown and it will fail at compile time.
On the other hand,

  T *t;
  #pragma omp task inout(t[1:2])
  { }

will work because, even if we do not know the exact type of 't', we
know it is a 'pointer to T' so by retaining this knowledge, OmpSs
transformation phases can still work correctly. A similar case happens
for arrays of dependent types.

This commit fixes #2311