Reuse analysis for the same function
These changes modify the vectorization/analysis strategy and
reuse the analysis among vectorization of loops within the same
function.

The key point here is to use the function 'register_identical_copy'
that allows to register code that was not originally present in the
code passed to the analysis.
This is only possible when new nodes have some identical equivalency
in the original code, such as main for loops and their respective epilogues.

In addition, because of the original loop and the 'pair' vectorized main loop +
epilogue loop are equivalent codes in terms of 'effects', it is
possible not to recompute analysis to vectorize other loops after
them.

See #2335