Copy the descriptor in a safer way
Passing a parameter of assumed shape array involves creating a descriptor,
if the real argument was already an assumed shape array, another copy
is created as well. This means that our current approach of capturing
the existing descriptor of an assumed shape array dummy argument
(i.e. a formal parameter that it is a dummy argument) is wrong since
we are memcpy-ing the copied descriptor address _after_ the function has
returned.

Note that this problem does not happen with deferred shape arrays since
for them a copy of a descriptor does not have to be created. The new code
works for these cases as well.

This commit fixes #2396