Add new flag --module-out-pattern
Since we want to use -JDIR to mean "create modules in directory DIR" we
must be able to tell the native compiler where to put its own modules.

Unfortunately, as it happens in the Fortran world, each compiler has its
own flag for that: gfortran uses "-JDIR", Intel Fortran uses "-module DIR"
and XL Fortran uses "-qmoddir=DIR".

To alleviate this the --module-out-pattern flag can be used to specify
a pattern to pas to the native compiler. By default it is "-J%s". The %s
placeholder will be expanded with DIR. Commas can be used to specify
more than one parameter. E.g, for Intel we need to use
--module-out-pattern="-module,%s".

Note that default of '-J%s' is intended for gfortran compatibility
out-of-the-box.

This commit fixes #2404

This commit fixes #