Calculate the locus of lists rather than materializing this
It is not obvious what is the locus of an ast that represents
a list. Until now it was the locus of the last item in the list.
This conflicts with the expected usage of ast_get_locus (and routines
like Nodecl::NodeclBase::get_locus) since one expects to be the locus
of the first item in the list.

Rather than materializing the locus, let it be empty and always calculate
it using the locus of the head of the list.

This commit fixes #2406