Reject declarations lacking a type-specifier
In C++ (not C) declarations cannot lack a type-specifier except for
constructors, conversion functions and destructors, otherwise the
declaration must contain a type.

The principal problem with successfully rejecting such cases is
identifying when the declaration names a constructor. In several places
we have to do some extra checks/lookups in order to be sure that we are
mentioning a constructor. The trick here is that a normal lookup (a
lookup not intended to find constructors) will find the injected
class-name. The injected class-name acts as a redundant public member of
a class, so its scope is that of the members of the class.

Other cases, i.e. destructors and conversions, are easily handled.  By
confining as much as possible the usage of the DF_CONSTRUCT flag we are
in another step towards removing this flag. The reason is that
constructors do not have proper names so their lookup is different
enough that regular lookup routines do not have to be used. This patch,
though, does not address this (questionable) design yet.

This patch removes build_scope_declarator_id_expr and moves its
behaviour into build_scope_declarator_name, since the latter was the
only caller of the former. Extra checks are performed in
build_scope_declarator_name to require (or reject) type-specifiers in
declarations.

Fixes #2410