Make sure failure tests verify the FE
Current failure tests just follow the usual approach of calling the
native compiler. This causes that we cannot distinguish a rejection in
the FE from a wrong codegen.

Some failure tests may still need to follow the whole pipeline if we
want them to ensure some behaviour based on the output (see
tests/02_typecalc_cxx.dg/failure_044.cpp).

Several friend declaration tests are now ignored since this is a
hair-pulling part of the standard that it is hard (and thankless)
and we do not really implement well.