Do not make codegen emit the definition of a class
When declaration/definition of a class member, codegen attempts
to define the class specifier. This is problematic when generating
new functions because it means it is very complex to add them
in the tree without triggering order issues.

Instead of that, just emit definitions of member symbols that
happened to be defined in non-class scope. For symbols that
were defined in class-scope (inside the class specifier) just
wait for the definition of the class to appear later.

See #2435