Use as a fallback the IEEE-754 Binary128
If the host compiler support __float128 then Mercurium supports it
everywhere.

While ideally we should verify the target supports it rather than the
host, we can fallback to a IEEE-754 Binary128 floating model in the case
the typing chosen typing environment does not.

This avoids a weird segmentation fault happening at early stages in
the compiler that prevents using of those typing environment (usually
for debugging purposes). Reconfiguring mercurium with --disable-float128
is feasible but unwieldy during debugging.

See #1698