Make multiexpressions a Mercurium extension
Currently multidependences are an extension for OmpSs but it seems
useful to use that syntax in other contexts.

Rather than using { ... } for C/C++ and ( ... ) for Fortran
this patch makes {/ ... /} the preferred syntax in both languages.

The old syntax is still allowed inside dependence clauses for
existing codes compatibility.

See #2448
See #2261