Analysis test phase propagates to graph nodes by default
Propagating information to graph nodes in the PCFG may be very costly.
In ticket #2335 we removed from the default behaviour this propagation
but it is needed sometimes. Thus, the analysis test phase requires
propagation by default and, those phases that ask for analysis (on demand)
do not have propagation by default.

It is needed to extend the analysis interface to allow asking for the
propagation when using the analysis on demand.

See #2335