Vectorize parameters during the vectorization of the function prototype
We split the vectorization of functions in two steps: the "declaration"
and their body. During the declaration phase we still changed the types
of local occurrences (in the body of the function).

This was ok as long as we did not check other vector functions. But
for pass-by-reference we need to see if the associated parameter has
been vectorized, so we need the information of vector functions be fully
computed prior checking the bodies of the functions.

Now we first vectorize parameters only during header vectorization and
we postpone vectorization of local variables for a later step.