Do not preprocess C++11 test
g++4.6 does not support -std=c++11 but we want this test to be run
in C++11 mode in Mercurium. The problem is that the file is first
preprocessed by passing -std=c++11 although it does not use any include
or macro.

If we do not preprocess it g++ is not called at all since this is a
test for an extension of Mercurium, where the native compiler is not
going to be called. We can skip preprocessing and thus avoid g++4.6
issues.

See #2489