Use !DEC$ ATTRIBUTE VALUE :: for Intel Fortran
Intel Fortran 15 changed the meaning of VALUE for non-interoperable
routines. We cannot switch to interoperable routiens in a reasonable
amount of time, so we have to resort on an Intel Fortran proprietary
extension directive to get the old behaviour.

Using `-assume nostd_value` is not an option since Intel Fortran 13 does
not understand that flag and would cause compilations to fail.

This commit fixes #2495