Fix list-initialization during overload
During overload, we may have te compute an ICS for a braced list
that involves initializing an aggregate type. This is a complex
algorithm already implemented in the expression typechecker, so we
use that one instead of repeating it.

We create a dummy expression for that, but initializer-lists are
not expressions so they need a special tree form them. We recursively
create it if needed.

This commit also fixes some laxity during list-initialization, where
we allowed excess initializers in contexts where them sould not be allowed.

This commit fixes #2509