Add support for non-local symbols
When non-local symbols are used, range analysis needs to introduce
a constraint, so each symbol has an initial value (the same happens
to parameters).
For constant non-local symbols, we catch the value of the symbol,
and for the rest, we introduce the [-inf, +inf] constraint.

Fixes #2517.

This commit also simplifies the way array accesses are represented.
So far we used the subscripts to create the SSA symbol. Since these
can be very complicated, now we use now the locus (with no path nor extension).