Better representation of subscripts in range analysis
In commit b214bce6b7b2b693cd9c72025aa24560dbbeba80 we introduced
a new way to represent subscripts using the locus. This is not
a good idea because we cannot match the the subscripts with their
ssa symbols.
This commit reimplements the initial version with better support
for complex subscripts.