Allow repetition of  constant nodes in the CG
In order to simplify the Constraint Graph, each value was represented
by a unique node. This is valid for nodes representing
SSA symbols, but it's not enough for constant values.

By reusing constant nodes, we were introducing undesirable cycles
wich lead to wrong a composition of the SCCs.

This commit enhances the representation, so constant nodes may appear
several times in the CG.