Overhaul the way we evaluate constants for constexpr
Currently we are setting the const value to the variable occurrence.
While this works right for values like scalars and tuples it does not
work to represent constant references to objects and subobjects.

C++11 concept of constant expression extends not only to rvalues
but also to glvalues (i.e. lvalues and xvalues) so we have to extend
const_value_t* with the concept of address and (sub)object.