Disable test
In its current form OpenMP::ForAppendix makes very hard to implement
its appendix part in the Nanos++ lowering.

For future reference: the appendix code (that does not have to only be
an epilog loop) happens to be in the same context as the main loop but
happens to be a sibling tree. During the Nanos++ transformation such
property is lost and temporaries and masks are not declared. In C++ it
looks like it works because of explicit declarations appearing in the
correct place but the symbol is still in the wrong scope.