When matching template specializations we incorrectly traverse the types
During a declaration of a function we may need to fix dependent typenames
so they can be properly compared.

The current function, though, incorrectly stops at specialized
template types. This makes sense for class types but it does not work
for function types (which should be handled as normal function types
at this point).

Since these types are just used for comparison purposes, make sure we
do not accidentally use them for anything else, and use a variable for
such purpose only.