Fix PCFG for return stmts generating a graph node
When the value of a return statement generates a graph node,
the return statement generates a 'split statement node'.
In this branch we were tagging as 'return node' the node generated
in a return statement for both cases, when a simple node or a
graph node are generated. For the former it is correct,
whereas for the latter it is not, because we are removing
its previous 'graph node' tag.

Now, when a split node is generated in a return statement, we look
for the real return statement and tag that one as the return node.

Example: return foo(p);
{{{
split_stmt
    entry
    func_call_graph
        entry
        foo(p)
        exit
    return foo(p)     --> This is the return node
    exit
}}}

With this fix, we re-enable a broken test