Restrict C initializations
In C the front end allows as standard conversions, conversions
between different pointer types (to avoid rejecting code that
traditionally has only been diagnosed as a warning).

On the other hand, C-style casts are classified according to C++ rules
but since C is more lax, cast between pointer types end being classified
as static_cast. This causes a wrong propagation of a constant
value (since only reinterpret_cast and dynamic_cast are not propagated).

Because of this, a cast to a pointer of vector type (from an lvalue
of array which has decayed to a pointer type) may cause the compiler
to attempt to convert a scalar to a vector in the wrong place.

This commit fixes #2583