Fix issue with Conversion and Dereference
The last changes related with the Converesion node
break the vectorization of Dereferences.
It seems that now there are new Conversion nodes not
expected by the vectorizer and the analysis.
This commit partially fixes the problem on the vectorizer
side.
There are several warnings on the analysis side.

A new test (nbody) has been added.