Perform partial dependent type synthesis on unary * and array subscript
C++ defines the concept of type dependent expressions that means
that the type of an expression may not be known at compile time
because it depends (maybe indirectly) on a template parameter.

During typechecking of expressions, most of the time it is not possible
to tell which is the type of an expression whose operands are type
dependent expressions. In this case, Mercurium synthesizes a node with
a "unknown dependent type".

That said, there are some simple cases, involving partially dependent
types, where the type is not completely unknown. Albeit the expression
is type dependent, it may have more precise typing information than
"unknown dependent type".

For instance

  template <typename T>
  void f(T** a)
  {
     a;       // T**& (already synthesized)
     a[1];    // T*&  (new, current is unknown dependent type)
     a[1][2]; // T&   (new, current is unknown dependent type)
  }

Note that it is very important to realize that the expressions are still
type-dependent in their nature but their associated type need not be
completely unknown.

This change may seem useful but enables doing extra transformations at the
template level for OmpSs programs. For instance, this is possible now

template <typename T>
void t(std::vector<T*> w)
{
    T** pw = w.data();
#pragma omp task inout({*(pw[i]), i = 0 : w.size() - 1})
    {
      ...
    }
}

(Note the usage of std::vector::data since we cannot still use
functions returning reference types)

This commit fixes #2584