Make Nanos++ again the default OmpSs runtime
When both Nanos 6 and Nanos++ were enabled in Mercurium,
Nanos 6 was preferred and --nanox had to be specified
for Nanos++.

This commit makes Nanos++ the default OmpSs runtime. This
is required for testing because we cannot change the default behaviour
of the git deployment.

Rather than changing all the existing installations to default to
Nanos++, it is easier to make Nanos++ the default and let the user
override this. The previous behaviour can be restored using

   --with-default-ompss-runtime=nanos6

at configure time.