Change the order in which compile versions are generated
We construct all compile versions by merging those in the test file
and those in the generator. The cartesian product takes first
the versions in the test file and then the versions of the generator.

It makes sense to me now to do it in the other way round: prioritize
those versions in the generator and then those in the test file. The reason
is that the test file may have compile-time dependences which the
generator is less likely to have.