Fix BIND specifier
When support for BIND(OPENCL, ...) was added the BIND-specifier stopped
being an AST_ATTR_SPEC, this means that we cannot handle it in a
homogeneous way like the other attributes.

This caused the problem described in #2599

Since we may want in the future support other BIND kinds other than C,
this patch introduces a new AST_BIND_KIND. Currently the implementation
only supports BIND(C, ...) and removes any existing BIND(OPENCL, ...)
handling.

This patch also simplifies the handling of BIND making it more similar
to the attr specifiers. It also fixes some functions to use a single
code path and avoid repetition.

This commit also removes NODECL_BIND_OPENCL as BIND(OPENCL, ...) is not
used anywhere in the compiler.

This commit fixes #2599