Make sure get_internal_nodecl does not allow accidental assignments
Nodecl::NodeclBase::get_internal_nodecl() returns a nodecl_t. Due to the
nature of the nodecl_t type, in C++ is possible to assign to a call like
this

 n.get_internal_nodecl() = ...;

But this does nothing to the tree because a copy of the nodecl_t
is returned. Probably in the past get_internal_nodecl() returned a
reference but we removed this to avoid problems, but we forgot to add
a const.

This commit fixes #2601