Removing useless input value dependences
This was a polemic feature that I added to Mercurium a few years ago. The
idea was to try to support input dependences over parameters passed by value in outlined tasks.

We decided to remove it since it was not really used and it complicated a lot the lowering phase.

Note that this commit breaks the backward  compatibility of our fortran modules.