Add parameter call_graph to avoid analyzing methods in the call graph
When parameter functions is used, all methods in the call graph of the functions in that list is also analyzed.
The parameter call_graph can be used to avoid analyzed functions in the call graph.
Its default value is 1 (meaning that functions in the call graph are analyzed).
Set it to 0 to disable the recursion as follows:
--variable=call_graph:"0"