An address constant expression may appear as a non-type template argument (2)
New implementation fixing this issue. Instead of adding more intelligence to the way how we
compare the constant values of the non-type argument expressions, what we do is skip this check for
constant values that are addresses or objects since we already have a code that handle them later.

It fixes  #2687