Fix error in PCFG tasks synchronization
The algorithm that determines tasks synchronizations iterates until no changes apply.
The last modifications (8545597c61bece82a3a43440280357384b9e92a9) made to this algorithm may change the synchronization kind of an edge after each pass of the algorithm. This breaks some assumptions made by the algorithm (a node may only appears once in the list of points of sync. If we do not check whether the node is there or not before inserting it, we may end up with the same node twice: one as static dependent, and the other as maybe dependent).

This commit fixes that.