Tasks synchronizations tunning was too naïve
The process of tuning tasks synchronizations is taking into account only the relation between a specific source with a specific target.
This is not enough, since the algorithm should take into account the relation of the source with other possible targets.
Example:
{{{
    out(a)  --maybe-->  in(a) // target1
           |--maybe-->  in(a) // target2
}}}
When tunning the relationship between the source and target1, we cannot change it from Maybe to Static, since the source can still synchronize with target2.
As a temporary fix, we will Keep the relationships as the task synchronization algorithm computes them. The task synchronization tunning process will only compute the conditions for each Maybe synchronization edge.
It remains as a future work fix this algorithm or remove the unnecessary tunning, if nothing to tune exists.

See #2691.