Propagate the DF_DEPENDENT_TYPENAME to lookups of nested names
Initially I was propagating this flag only when the current symbol
(i.e. the symbol associated with the current name) was in the lexical scope
(otherwise we already created a new dependent entity). However, since it seems
that I doesn't break anything, I decided to always propagate it. If at some
point we find out that this may be problematic, we can always constrain this propagation.

It fixes #2714