Reduce memory usage during TDG expansion
Nodecl structure is used to compare dependency expressions.
For each pair of tasks that may be dependent, the expression is checked with the proper values.
This requires to perform a shallow_copy of the expression. This new nodecl_t must be freed. Otherwise, the algorithm runs out of memory very soon.