Fix liveness for OpenMP nodes
Liveness analysis must take into account the visibility of the variables in OpenMP nodes.
So far, we were treating specially only task and target constructs, but all constructs defining specific visibility must be considered.
Furthermore, all visibilities (shared, firstprivate, lastprivate and private) must be checked. And they have to be considered for both Live In and Live Out sets computation.

This commit also adds a test for this feature.