Fix tasks synchronizations mechanism
In 5cc02af01c090dc7786c9a144099a9b248938c1b we broke the algorithm because we overloaded the result "unknown" in method "compute_task_sync_relationship". We were using the value unknown to specify:
* A dependence may exist here
* A dependence exists but the source can still cause other dependences

Now, we use unknown for what it is meant (the dependence may exist), and we use an extra parameter to define the cases where we want the source task to be alive even though it has been synchronized.