Make C_NULL_PTR a constant and fix a problem with array initialization
When making C_NULL_PTR a constant we discovered that we assume it will
be possible to update the expression of a scalar that initializes an
array to a tree derived from the constant expression.

Don't do this as it is not always possible (we are not doing for
arrays whose size is not known so we shouldn't be losing much).