Implement deferred character length using a descriptor array
This is only useful for CHARACTER(LEN=:)

In Fortran we implement CHARACTER scalar type using a C array of
characters. That array, which is not considered a Fortran array,
never used to have descriptor. Now that LEN can be deferred, we
will represent this using a descriptor at that array.