Add new implementation of TL::ForStatement utility
Despite that this new implementation supports more cases than the previous one
(e.g. for (unsigned int x = 0; x < 0u; ++x) {} ) it's not enabled by default.

The main problem is that we are changing the behavior of an utility that is used
in several places. Ideally at some point we should be able to deprecate the old behavior of this utility.


It fixes #2796