Do not keep a fake dependent type in expressions
When a qualified-id mentions a dependent entity in an expression, like

  A<T>::value

we used to associate the expression with a dependent typename. So the
type of this expressions was equivalent to

  typename A<T>::value

but this does not make sense because if A<T>::value is valid in an
expression, it will never be valid for a type. That dependent typename
is only useful for the contexts where we do care about dependent
typenames, like

  typename A<T>::value Q;

This leads to confusion to other parts of the compiler, so let's stop
doing this.

This change should not have any functional change because an unknown
dependent type is still a dependent type as it was a dependent typename.