Do not allow dependent typenames that and referring to variables
In 9289b5dd338609a1428dd74752b6ab1d89aa102b we changed this function
so when after updating the dependent typename we end in an enum, we
just return the type of the variable.

Unfortunately this is incorrect because it means that "typename T::Q"
where "T::Q" refers a variable is now valid.

This bug was introduced while trying to fix another bug where we
aggressively try to update the type of non template-arguments. Usually
the type of nontype template-arguments can be matched to the type of the
corresponding non type template-parameter. Unfortunately when the
template-argument list contains an expansion we refrain ourselves from
doing this because we may not be able to match arguments with parameters
when we parse the template arguments. When this happens the type of the
argument is just the type of the expression.

We used to associate to qualified-id's of dependent expressions a silly
dependent typename type. As we stopped doing this, the update will not
be attempted. When this happens, we just use the type of the expression.